+++ example/convex/_generated/server.d.ts
@@ -150,124 +150,4 @@ export type DatabaseReader = GenericDatabaseReader<DataModel>;
  */
 export type DatabaseWriter = GenericDatabaseWriter<DataModel>;
 
-export declare const components: {
-  pushNotifications: {
-    public: {
-      deleteNotificationsForUser: FunctionReference<
-        "mutation",
-        "internal",
-        { logLevel: "DEBUG" | "INFO" | "WARN" | "ERROR"; userId: string },
-        any
-      >;
-      getNotification: FunctionReference<
-        "query",
-        "internal",
-        { id: string; logLevel: "DEBUG" | "INFO" | "WARN" | "ERROR" },
-        null | {
-          body?: string;
-          data?: any;
-          numPreviousFailures: number;
-          sound?: string;
-          state:
-            | "awaiting_delivery"
-            | "in_progress"
-            | "delivered"
-            | "needs_retry"
-            | "failed"
-            | "maybe_delivered"
-            | "unable_to_deliver";
-          title: string;
-        }
-      >;
-      getNotificationsForUser: FunctionReference<
-        "query",
-        "internal",
-        {
-          limit?: number;
-          logLevel: "DEBUG" | "INFO" | "WARN" | "ERROR";
-          userId: string;
-        },
-        Array<{
-          body?: string;
-          data?: any;
-          id: string;
-          numPreviousFailures: number;
-          sound?: string;
-          state:
-            | "awaiting_delivery"
-            | "in_progress"
-            | "delivered"
-            | "needs_retry"
-            | "failed"
-            | "maybe_delivered"
-            | "unable_to_deliver";
-          title: string;
-        }>
-      >;
-      getStatusForUser: FunctionReference<
-        "query",
-        "internal",
-        { logLevel: "DEBUG" | "INFO" | "WARN" | "ERROR"; userId: string },
-        { hasToken: boolean; paused: boolean }
-      >;
-      pauseNotificationsForUser: FunctionReference<
-        "mutation",
-        "internal",
-        { logLevel: "DEBUG" | "INFO" | "WARN" | "ERROR"; userId: string },
-        null
-      >;
-      recordPushNotificationToken: FunctionReference<
-        "mutation",
-        "internal",
-        {
-          logLevel: "DEBUG" | "INFO" | "WARN" | "ERROR";
-          pushToken: string;
-          userId: string;
-        },
-        null
-      >;
-      removePushNotificationToken: FunctionReference<
-        "mutation",
-        "internal",
-        { logLevel: "DEBUG" | "INFO" | "WARN" | "ERROR"; userId: string },
-        null
-      >;
-      restart: FunctionReference<
-        "mutation",
-        "internal",
-        { logLevel: "DEBUG" | "INFO" | "WARN" | "ERROR" },
-        boolean
-      >;
-      sendPushNotification: FunctionReference<
-        "mutation",
-        "internal",
-        {
-          allowUnregisteredTokens?: boolean;
-          logLevel: "DEBUG" | "INFO" | "WARN" | "ERROR";
-          notification: {
-            body?: string;
-            data?: any;
-            sound?: string;
-            title: string;
-          };
-          userId: string;
-        },
-        string | null
-      >;
-      shutdown: FunctionReference<
-        "mutation",
-        "internal",
-        { logLevel: "DEBUG" | "INFO" | "WARN" | "ERROR" },
-        { data?: any; message: string }
-      >;
-      unpauseNotificationsForUser: FunctionReference<
-        "mutation",
-        "internal",
-        { logLevel: "DEBUG" | "INFO" | "WARN" | "ERROR"; userId: string },
-        null
-      >;
-    };
-  };
-};
-
 /* prettier-ignore-end */
