+++ example/convex/convex.config.ts
@@ -1,5 +1,5 @@
 import { defineApp } from "convex/server";
-import pushNotifications from "../../src/component/convex.config.js";
+import pushNotifications from "../../src/component/convex.config";
 
 const app = defineApp();
 app.use(pushNotifications);
