+++ example/package.json
@@ -5,7 +5,7 @@
   "version": "0.0.0",
   "main": "app/App.tsx",
   "scripts": {
-    "dev": "convex dev --live-component-sources",
+    "dev": "convex dev --live-component-sources --typecheck-components",
     "dev:frontend": "vite",
     "logs": "convex logs",
     "lint": "tsc -p convex && eslint convex"
@@ -13,7 +13,7 @@
   "dependencies": {
     "@convex-dev/expo-push-notifications": "file:..",
     "@tamagui/config": "^1.110.3",
-    "convex": "1.16.0-alpha.2",
+    "convex": "^1.16.5",
     "expo": "^51.0.32",
     "expo-constants": "^16.0.2",
     "expo-device": "^6.0.2",
