+++ src/client/index.ts
@@ -6,6 +6,9 @@ import {
   GenericQueryCtx,
 } from "convex/server";
 import { GenericId } from "convex/values";
+import { api } from "../component/_generated/api.js";
+import { NotificationFields } from "../component/schema.js";
+import { LogLevel } from "../logging/index.js";
 
 /**
  * This component uses Expo's push notification API
@@ -196,11 +199,6 @@ type RunQueryCtx = {
 type RunMutationCtx = {
   runMutation: GenericMutationCtx<GenericDataModel>["runMutation"];
 };
-
-// TODO: Copy in a concrete API from example/_generated/server.d.ts once your API is stable.
-import { api } from "../component/_generated/api.js"; // the component's public api
-import { NotificationFields } from "../component/schema.js";
-import { LogLevel } from "../logging/index.js";
 
 export type OpaqueIds<T> =
   T extends GenericId<infer _T>
