+++ src/component/_generated/api.d.ts
@@ -165,4 +165,6 @@ export declare const internal: FilterApi<
   FunctionReference<any, "internal">
 >;
 
+export declare const components: {};
+
 /* prettier-ignore-end */
